requirements.txt
flake8
black
fastapi==0.75.1
uvicorn
python-ffmpeg-video-streaming